Sacado:  Move new-design function declarations to base file.
